[Feature] Add Inductive Miner changes in inductive_miner directory. (#326)
- [ ] BugFix
- [X] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related
issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

<!-- (Please describe the changes you have made. In which scenarios will
this bug be triggered and what measures have you taken to fix the bug?)
-->
Add Inductive Miner changes in inductive_miner directory. There will be
a separate PR or two for the rest depending on the progress of PRs for
upstream.

CELOSTAR is used for conditional compilation.
I tried to make things merge friendly by making all diffs as additions.
These changes would be to able to be applied to cpm-query-enqine
repository. But how to share codes will be a separate but important
topic in the near future.

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

- [ ] I have checked the version labels which the pr will be auto
backported to target branch
  - [ ] 2.5
  - [ ] 2.4
  - [ ] 2.3
  - [ ] 2.2