[BugFix][Feature] Implement convert_to_serialize_format() to support streaming aggregation of celonis_sorted_first(). (#623)
Previously, convert_to_serialize_format() was not implemented and only
had DCHECK() to warn this.
But in production, when a query was executed as a streaming aggregation,
the function was called, DCHECK() was skipped and it caused crashes
later in exchange.

For other aggregation functions potentially with the same issue, this PR
converted DCHECKs to throw std::runtime_error. In production, it will
return an error with a message instead of causing crashes.

I tested this locally. I'll merge it after testing the image being built
with a large data set successfully.

## What type of PR is this:
- [X] BugFix
- [X] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [X] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4

---------

Co-authored-by: Hector Gonzalez <hagonzal@users.noreply.github.com>