[Feature] Add celonis_calc_bucket_boundaries(). (#665)
Outputs a boundary array to be used as an input of
celonis_histogram_boundaries() to implement PQL HISTOGRAM with mode
BUCKET_COUNT.
https://celonis-confluence.atlassian.net/wiki/spaces/PQLdevelopment/pages/11245736/HISTOGRAM

Saola aggregates all data and trims top 5% and bottom 5% to get a bucket
width and adjusts boundaries later.
To support large scale data, we use approximate percentile to find
boundaries. So the boundaries may be different than Saola's.

This requires 2 AF calls to build a histogram with bucket_count mode.
Ideally, we need to add an AF which builds histogram in one pass.

```
celonis_calc_bucket_boundaries(col, count);
```
 * col : BIGINT | DOUBLE | DATETIME
 * count: BIGINT. Optional. Default is 10.
 * return: ARRAY of col type

- [ ] BugFix
- [X] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

- [X] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4