[BugFix] Registration order of celonis_calc_bucket_count_boundaries() for multiple arguments. (#668)
When registering a function for different types, it's been a practice to
register them in order of BIGINT, DOUBLE, DATETIME and VARCHAR.

It works well for strict matches but it turned out that it didn't work
for cast matches.

Before the PR, the function with count argument was registered as

1. Type.BIGINT, Type.BIGINT
2. Type.DOUBLE, Type.BIGINT
3. Type.DATETIME, Type.BIGINT

When it was called as
`celonis_calc_bucket_count_boundaries(datetime_col, 10)`, the argument
type is mapped to `Type.DATETIME, Type.TINYINT`. Because there is no
strict match, FE searches for cast matches and matches `Type.BIGINT,
Type.BIGINT` first.

So the PR changes the order like below. `Type.DATETIME` should be before
`Type.BIGINT` and for the same reason, `Type.BIGINT` should be before
`Type.DOUBLE`.

1. Type.DATETIME, Type.BIGINT
2. Type.BIGINT, Type.BIGINT
3. Type.DOUBLE, Type.BIGINT

## What type of PR is this:
- [X] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4