[BugFix] Update the registration order of some UDF functions in functions.py (#725)
Put the ARRAY_VARCHAR overload to be the last.

This is because SR optimizer iterates over the functions with the same
name in the order specified by functions.py, and picks the first one
that can be used (based on type casting). Most of the types can be cast
to VARCHAR.

Tested locally.

- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [x] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
- [ ] I have added documentation for my new feature or new function

- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4