fix: Fix issue in product operator where null column was not resized for internal fields of struct column (#777)
In convert_to_serialize_format we basically write into a nullable struct
column with two fields which each are nullable columns again. When we
resized the column, before we only resized the data columns of the
nullable struct field columns and the null columns of the nullable
struct column, but not the null columns of the individual nullable
struct field columns. This PR dfixes this by just resizing the entire
struct column rather than only the data columns of the individual
fields.

Fixes #issue

## What type of PR is this:
- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
- [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4