fix: Make sure null columns are correctly written / appended in the product aggregation operator (#784)
Make sure that nulls are also appended when appending to struct columns
fields. This is done by just calling append_datum on the struct column
(which should automatically append nulls as well) rather than appending
to the data columns manually.

Also adjusted the tests so that the column structure is consistent with
the one used in reality. Also called check_or_die sanitity checks on
columns in tests, so that tests fail when columns are in an inconsistent
state (e.g. data and null columns dont match)

## What type of PR is this:
- [ x ] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [ x ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
- [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4