[Feature] Support edge_count in celonis_variant_stats(). (#791)
`edge_count`: Limits the size of the edge table. It returns the first
`edge_count` edges it sees (so not ordered specifically).

As SR does not handle constant columns correctly in merge stage, we need
to pass optional constants through serialization. For that reason, the
existing inductive miner constant passing was fragile and needed to be
updated as well. It required some refactoring.

### Syntax
```
celonis_variant_stats(input_column, weight_column [, edge_count])
```

## What type of PR is this:
- [ ] BugFix
- [X] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [X] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
- [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4