[Enhancement] Improve CELONIS_ADD_TIMEUNITS_CALENDAR (#1035)
The original implementation of add_timeunits_calendar scans all the time
ranges in the Calendar. The time complexity is O(N) where N is the
number of time ranges. The new implementation uses binary search and the
time complexity is O(log(N)). The new algorithm only works when the time
ranges are non-weekly and time_unit != `WORKDAYS`.

Tested locally.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5