[Enhancement] Do not check neighbor with isolated nodes in DBSCAN (CLUSTER_VARIANTS) (#1080)
This optimization is based on the following observation: if a node is
labelled as isolated (does not connect with any other nodes) in previous
iterations, it can not be neighbors of any other nodes. Note that a
isolated node can be a core node.

For Ocado's data, it reduces the DBSCAN time from 12.5 s to 8.8 s on M2
MacBook Pro. It reduces the DBSCAN time from 20.5 s to 14 s on EC2.

It reduces the Ocado query time from 1 min 5 s to 48 s (DBSCAN time is
about 29 s).

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5