Retry the whole rebase step if it fails (#1188)
Based on previous observation, if a rebase step fails, it could be
- there are conflicts which the workflow could not handle and we need to
resolve manually
- `git rebase` is flaky

Previously it retried within the same step, but it did not work.
Recently I noticed that in the second situation rerunning the step would
resolve it
([example](https://github.com/celonis/celostar-starrocks/actions/runs/10365106107/job/28723246394))

This PR is to implement it by retrying the whole step for 3 times.

Manually tested and it looks good. Note there could be some
discrepancies between local and main branches. Will keep an eye on it
once it merges.