Revert the rebase change (#1191)
In the previous https://github.com/celonis/celostar-starrocks/pull/1188
and https://github.com/celonis/celostar-starrocks/pull/1189, I tried to
add retry logic for the whole rebase step to handle the flakiness of
rebase. The flakiness is that if I rerun manually for a failed rebase,
it sometimes succeeded (see the
[example](https://github.com/celonis/celostar-starrocks/actions/runs/10365106107/job/28723246394)).
During the testing the PR
https://github.com/celonis/celostar-starrocks/pull/1189, it worked
locally (see the test link in the PR's description), but it failed after
the PR merged
([example](https://github.com/celonis/celostar-starrocks/actions/runs/10381065008)).
Revert the change to unblock. Need more investigation on this issue.

Test locally:
-
https://github.com/celonis/celostar-starrocks/actions/runs/10382419797/job/28745470361
(The first step `sync-and-rebase-upstream` passed)