[Enhancement] Improve the efficiency of CELONIS_MATCH_ACTIVITIES (ENDING_NODES) (#1289)
Change the implementation to traverse the activity array according to
STARTING_NODES and ENDING_NODES.
1. STARTING_NODES is empty and ENDING_NODES is empty: traverse from left
to right.
2. STARTING_NODES is not empty and ENDING_NODES is empty: traverse from
left to right.
3. STARTING_NODES is empty and ENDING_NODES is not empty: traverse from
right to left.
4. STARTING_NODES is not empty and ENDING_NODES is not empty: if
STARTING_NODES.size() <= ENDING_NODES.size(), traverse from left to
right, otherwise traverse from right to left.

This improvement does not change the efficiency of other modes. For the
case that ENDING_NODES is set, it reduces the latency 50-70%.

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5