[BugFix] Set precision=17 when converting double to string in CELONIS_BUILD(_MULTI)_LINEAR_REGRESSION_MODEL (#1407)
In the current implementation when generating model string, we use
precision = 6. This can lead to large error when operating on unix
milliseconds space.

For example in the following example ([repro
SQL](https://gist.github.com/zhangyaqi1989/bed76833033868df185ba202cc9cbb7f)),
the model should be
`-164992598.46075037121772766113:0.00010083605724405435`
If we use precision = 6, the model is
`-164992598.460750:0.000101`
When doing prediction on x = 1640995200000. The predicted_y =
747916.7392500043. However using high precision model, the predicted_y
is 478887.4634427726.

This PR changes to use precision = 20.

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5