[Enhancement] Improve CELONIS_MATCH_STRINGS further (handle const top_k and separator in prepare) (#1432)
## Why I'm doing:

In the original implementation of CELONIS_MATCH_STRINGS, we only check
if the match_array column is constant or not. Since PQL2SQL is the only
user of this function, and currently both top_k and separator columns
are constant. This PR changes to also process top_k and separator
columns in prepare().

This PR also switched to use SliceWithHash instead of string as key in
the cache.

This is a follow up PR of
https://github.com/celonis/celostar-starrocks/pull/1422.

## What I'm doing:

Based on the benchmark, this change improves the performance of const
version of CELONIS_MATCH_STRINGS further.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5