[BugFix] Fix a bug in CELONIS_MERGE_SORTED_ARRAYS (may not order null sorting before non-null sortings) (#1451)
## Why I'm doing:

[PR](https://github.com/celonis/celostar-starrocks/pull/1437) introduces
a bug which makes CELONIS_MERGE_SORTED_ARRAYS may not order NULL sorting
before non-NULL sortings. This is caused by trying to access NULL
element through `elements` of the array (sorting) column. The NULL
element in `elements` is not guaranteed to be set to KNullDatum. Note
that this bug does not cause crash and it is not deployed yet.

## What I'm doing:

This PR changes to check if a sorting element is NULL through
`null_elements`.

This PR also adds a unit test to double check the behavior. The original
unit test failed to catch this bug during merging the original PR (it
becomes to a flaky test).

This PR also reruns the benchmark, the change is ignorable.

## What type of PR is this:

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5