CELOSTAR-357 Add celonis_array_lag_window (#1517)
## Why I'm doing:

Slack discussion in #celostar-dev for context:
https://celonis.slack.com/archives/C06FNK5TNQY/p1738333199909749

PQL has a LAG operator that can be used without PARTITION BY. For
Celostar this means that the entire table needs to be sorted in the
window framework which usually leads to the BE reporting `Total byte
size of binary column exceed the limit: 4294967296: BE:11016` and the
query fails.

In case the table is an activity table, we currently unnest the arrays
on the case table and then perform the window lag operator. The idea is
to implement a custom array_lag_window function that performs the LAG
operation directly on the arrays in the case table instead of unnesting.

## What I'm doing:

This PR adds `celonis_array_lag_window(array, k)`.

- array can be any celonis typed array
- k is the offset parameter. Must be a positive (>0) constant.
- The function returns an array of the same type as the first arg.

Note that in contrast to other window functions to windowing clause can
be specified. This is equivalent to how LEAD/LAG behaves in SR.

Here is an example:

```
mysql> select id, col1, col2, celonis_array_lag_window(col2, 1) over (partition by col1 order by id) from t order by col1, id;
+------+------+-------------------------+-----------------------------------------------------------------------------+
| id   | col1 | col2                    | celonis_array_lag_window(col2, 1) OVER (PARTITION BY col1 ORDER BY id ASC ) |
+------+------+-------------------------+-----------------------------------------------------------------------------+
|    1 |    4 | [4]                     | [null]                                                                      |
|    3 |    4 | [4,5,6]                 | [4,4,5]                                                                     |
|    4 |    4 | [null,1,2,3]            | [6,6,1,2]                                                                   |
|    5 |    4 | [10,null,11,12,null,13] | [3,10,10,11,12,12]                                                          |
|    2 |   10 | [10]                    | [null]                                                                      |
+------+------+-------------------------+-----------------------------------------------------------------------------+

```

## What type of PR is this:

- [ ] BugFix
- [x] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5