CELOSTAR-397: Add weekday calendar benchmark (#1546)
## Why I'm doing:

After invesetigating a slow query for Poste Italiane Spa, we found that
the query was slow due to the weekday calendar used.

## What I'm doing:

Adds a benchmark for the weekday calendar. Right now it looks like the
weekday calendar is about 8x slower than the workday calendar. The
operator will be optimized in a follow-up.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5