CPL-13820 Add generic utilities to enable self-contained top k sorting of arbitrary rows (#1601)
## Why I'm doing:

This is required for moving_top_k operator, see [this
thread](https://celonis.slack.com/archives/C06FNK5TNQY/p1733418303714869?thread_ts=1733242713.862379&cid=C06FNK5TNQY)
for context. The issue with implementing such a window aggregation is
that the state needs to keep track of the current top k rows - as the
state might survive the existence of individual frame chunks this
requires not just remembering the indices but also the row contents,
which is not trivial as rows can have arbitrary type signatures/

## What I'm doing:

This PR does essentially four things (and it makes sense to review them
in this order):
* Add functionality to compare arbitrary Datum instances, using
additional config like SortOrder and NullHandling (see sort_utils.h)
* Extend serialization utils to work with all types TimestampValue,
DateValue and DecimalValue. Also adds a function that verifies whether a
type is statically serializable and adds a constexpr function for
computing the serialized size when that is the case
* Implement a row buffer clas. This is a generic framework for storing
and serializing a set of rows of arbitrary variadic type signatures into
a fixed memory chunk. It analyzes the type signature and constructs the
full memory chunk on construction and then offers slots to set/read
individual rows. Most types are handled by straightforward
serialization. Strings are more complicate and are handled by
maintaining an dynamically allocated and updated string buffer that the
serialized memory chunk points to.
* Implement a TopKSorter class (top_k_utils.h). This is a class which
essentially abstracts a top-k priority queue while remembering the
values of the current top-k rows. For this it internally uses a row
buffer. Since the row buffer operates on fixed memory, the TopKSorter
can efficiently overwrite and remember the current topK rows without
needing to reallocate memory (except in the case of strings).

For all four things I added or extended tests

The implementation of the moving top_k operator can use these generic
utilities and will then be relatively straightforward

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [ x ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [] Yes, this PR will result in a change in behavior.
- [X] No, this PR will not result in a change in behavior.

## Checklist:

- [X] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5